refactor: nodes extraction
