+++ cognee/tasks/graph/extract_nodes_from_data.py
@@ -1,13 +1,15 @@
 import asyncio
 from typing import List
+from pydantic import BaseModel
 
 from cognee.infrastructure.databases.graph import get_graph_engine
 from cognee.modules.chunking.models.DocumentChunk import DocumentChunk
 from cognee.modules.data.extraction.knowledge_graph.extract_content_graph import (
-    extract_content_nodes,
     EntityListResponse,
 )
 from cognee.tasks.storage import add_data_points
+from cognee.infrastructure.llm.get_llm_client import get_llm_client
+from cognee.infrastructure.llm.prompts import render_prompt
 
 
 async def extract_nodes_from_data(
@@ -23,3 +25,36 @@ async def extract_nodes_from_data(
         chunk.potential_nodes = nodes
 
     return data_chunks
+
+
+class PotentialNodes(BaseModel):
+    """Response model containing a list of potential node names."""
+
+    nodes: List[str]
+
+
+async def extract_content_nodes(content: str, n_rounds: int = 2) -> List[str]:
+    """Extracts node names from content through multiple rounds of analysis."""
+    llm_client = get_llm_client()
+    all_nodes: List[str] = []
+    existing_nodes = set()  # Track existing node names in lowercase
+
+    for round_num in range(n_rounds):
+        context = {
+            "previous_nodes": all_nodes,
+            "round_number": round_num + 1,
+            "total_rounds": n_rounds,
+        }
+
+        system_prompt = render_prompt("extract_graph_nodes_prompt.txt", context)
+        response = await llm_client.acreate_structured_output(
+            text_input=content, system_prompt=system_prompt, response_model=PotentialNodes
+        )
+
+        # Only add new nodes that haven't been seen before
+        for node in response.nodes:
+            if node.lower() not in existing_nodes:
+                all_nodes.append(node)
+                existing_nodes.add(node.lower())
+
+    return all_nodes
